fix: extract and update calculateTimeRemaining utility for localization
Refactors the time remaining calculation into a shared utility function, ensuring consistent localized formatting across components and services, and removes duplicate implementations to improve maintainability.
