+++ src/services/notification.service.ts
@@ -25,6 +25,7 @@ import {
 } from '../constants/verifiedEmail.constant'
 import { contactSummaryText, contactFullText, saleAlertText } from '../constants/discord.constant'
 import { getTranslations } from '../lib/i18n/dictionary'
+import { calculateTimeRemaining } from '../lib/utils/date'
 
 const renderContactMessageEmailHtml = (p: { name: string; email: string; message: string }): string =>
 	(renderContactMessageEmailHtmlUnsafe as unknown as (p: { name: string; email: string; message: string }) => string)(p)
@@ -111,78 +112,7 @@ export async function sendNewBibNotification(
 		return { success: true, emailsSent: 0, emailsFailed: 0, adminNotified: false }
 	}
 
-	// Calculate time remaining until event
-	const calculateTimeRemaining = (eventDate?: Date | string, locale: string = 'fr'): string => {
-		try {
-			// Use translations to get the time units following the same pattern as components
-			const translations = getTranslations(locale, {})
-
-			// Type-safe access to timeUnits with fallback
-			const timeUnits = translations.GLOBAL?.timeUnits as
-				| {
-						defaultDays: string
-						zeroDays: string
-						oneDay: string
-						days: string
-						oneWeek: string
-						weeks: string
-						oneMonth: string
-						months: string
-				  }
-				| undefined
-
-			// Fallback if timeUnits are not available
-			if (!timeUnits) {
-				const fallbackDays = locale === 'en' ? '7 days' : '7 jours'
-				if (eventDate === undefined || eventDate === null || (typeof eventDate === 'string' && eventDate.trim() === ''))
-					return fallbackDays
-
-				try {
-					const event = new Date(eventDate)
-					const now = new Date()
-					const diffTime = event.getTime() - now.getTime()
-					const diffDays = Math.ceil(diffTime / (1000 * 60 * 60 * 24))
-
-					if (diffDays <= 0) return locale === 'en' ? '0 day' : '0 jour'
-					if (diffDays === 1) return locale === 'en' ? '1 day' : '1 jour'
-					if (diffDays < 7) return locale === 'en' ? `${diffDays} days` : `${diffDays} jours`
-					const weeks = Math.floor(diffDays / 7)
-					if (weeks === 1) return locale === 'en' ? '1 week' : '1 semaine'
-					if (weeks < 4) return locale === 'en' ? `${weeks} weeks` : `${weeks} semaines`
-					const months = Math.floor(diffDays / 30)
-					return months === 1
-						? locale === 'en'
-							? '1 month'
-							: '1 mois'
-						: locale === 'en'
-							? `${months} months`
-							: `${months} mois`
-				} catch {
-					return fallbackDays
-				}
-			}
-
-			if (eventDate === undefined || eventDate === null || (typeof eventDate === 'string' && eventDate.trim() === ''))
-				return timeUnits.defaultDays
-
-			const event = new Date(eventDate)
-			const now = new Date()
-			const diffTime = event.getTime() - now.getTime()
-			const diffDays = Math.ceil(diffTime / (1000 * 60 * 60 * 24))
-
-			if (diffDays <= 0) return timeUnits.zeroDays
-			if (diffDays === 1) return timeUnits.oneDay
-			if (diffDays < 7) return `${diffDays} ${timeUnits.days}`
-			const weeks = Math.floor(diffDays / 7)
-			if (weeks === 1) return timeUnits.oneWeek
-			if (weeks < 4) return `${weeks} ${timeUnits.weeks}`
-			const months = Math.floor(diffDays / 30)
-			return months === 1 ? timeUnits.oneMonth : `${months} ${timeUnits.months}`
-		} catch {
-			// Ultimate fallback
-			return locale === 'en' ? '7 days' : '7 jours'
-		}
-	}
+	// Calculate time remaining until event (now imported from date utils)
 
 	const formatEventDate = (date?: Date | string) => {
 		if (date === undefined || date === null || (typeof date === 'string' && date.trim() === '')) return ''
